MyBackend/requirements.txt
arabic-reshaper==3.0.0
asgiref==3.7.2
asn1crypto==1.5.1
astroid==3.1.0
blinker==1.8.2
Brotli==1.1.0
certifi==2024.2.2
cffi==1.16.0
channels==4.1.0
channels-redis==4.2.0
chardet==5.2.0
charset-normalizer==3.3.2
choices==0.1
click==8.1.7
colorama==0.4.6
contourpy==1.2.1
cryptography==42.0.5
cssselect2==0.7.0
cycler==0.12.1
defusedxml==0.7.1
diff-match-patch==20230430
dill==0.3.8
distlib==0.3.9
Django==5.0.3
django-cors-headers==4.5.0
django-import-export==3.3.8
django-widget-tweaks==1.5.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
et-xmlfile==1.1.0
filelock==3.16.1
Flask==3.0.3
fonttools==4.51.0
html5lib==1.1
idna==3.7
iniconfig==2.0.0
isort==5.13.2
itsdangerous==2.2.0
Jinja2==3.1.4
kiwisolver==1.4.5
libmagic==1.0
lxml==5.2.1
MarkupPy==1.14
MarkupSafe==2.1.5
matplotlib==3.9.0
mccabe==0.7.0
msgpack==1.0.8
mysqlclient==2.2.4
numpy==1.26.4
odfpy==1.4.1
openpyxl==3.1.2
oscrypto==1.3.0
packaging==24.0
pandas==2.2.2
pillow==10.2.0
pipenv==2024.2.0
platformdirs==4.2.1
plotly==5.22.0
pluggy==1.5.0
pycparser==2.22
pydyf==0.11.0
pyHanko==0.23.2
pyhanko-certvalidator==0.26.3
PyJWT==2.9.0
pylint==3.1.0
pyparsing==3.1.2
pypdf==4.2.0
PyPDF2==3.0.1
pyphen==0.14.0
pypng==0.20220715.0
pytest==8.2.1
python-bidi==0.4.2
python-dateutil==2.9.0.post0
python-magic==0.4.27
python-magic-bin==0.4.14
pytz==2024.1
PyYAML==6.0.1
qrcode==7.4.2
redis==5.0.7
reportlab==4.0.9
requests==2.31.0
runserver==1.0.4
setuptools==75.2.0
six==1.16.0
sqlparse==0.4.4
svglib==1.5.1
tablib==3.5.0
tenacity==8.3.0
tinycss2==1.3.0
tomlkit==0.12.4
typing_extensions==4.11.0
tzdata==2024.1
tzlocal==5.2
uritools==4.0.2
urllib3==2.2.1
virtualenv==20.27.0
weasyprint==62.3
webencodings==0.5.1
Werkzeug==3.0.3
xhtml2pdf==0.2.15
xlrd==2.0.1
xlwt==1.3.0
zopfli==0.2.3
